fix: Serve move-modal-action assets only when the action is available
request #33651 move-artifact-action app should be properly unit tested

Since we now use @tuleap/dom [0], an error is thrown when the mount point
can't be found. Given that the assets were served even if there is no
move artifact mount point in the DOM, an error is poping in the
console when the user is not tracker administrator.

[0] See git #tuleap/stable/0568dff24d6c854517a0c74c503f8688e12f0744

How to test:
- As a regular user, open an artifact
--> There is no error in the console about #move-artifact-modal not
being found in the DOM

- As a tracker admin, you can move an artifact using the modal as
  before.

Change-Id: Ie8a05e0dbe89db74f0eefcb92c55bdc4f853a8c6