Move the kanban package.json used for E2E tests to the expected path
There is no need to have it at the root of the plugin. This is a source
of issues and headaches when there are packages in the subfolders due to
the way the dependencies relation for the cache layer are built.

Part of story #33646: access Kanban in a dedicated service

Change-Id: Iccd847c4cf78ff2d0c7cb5aa1dfd04a0a795a077