feat: Move export of kanban in kanban
Absence/presence of A.D plugin should not impact Import/Export of
kanban.

This duplicates definition of kanban list. The one in A.D is used during
project import to validate the imported XML. The one in kanban is used
during export to validate the exported kanban configuration. Currently I
have no idea on how to ignore A.D stuff in kanban (and vice et versa),
and how to validate the whole during project import…

A.D rnc is updated to:
* ignore order of children of `<agiledashboard>` (kanban list can be
  added before A.D)
* allow `<agiledashboard>` to not have any `<planning>` element (when
  A.D is deactivated, there is no exported `<planning>`)

Note: Kanban widget does not export itself when it refers to a non existing
kanban.

Part of story #33646: access Kanban in a dedicated service

Change-Id: Iddadd7b97439cead1e498cc55f35545d24ee023e