chore: rename src/common/log to src/common/Log
This is part of request #33669 Logger level can be set using the Tuleap CLI

No functional change expected here.
This is done to match the namespace Tuleap\Log
$level parameter in TruncateLevelLogger has been set to mixed to avoid
to rewrite a lot of code because of ForgeConfig::get() can return false
and is not set in a lot af unit tests.

Change-Id: Id9e5d156450380c9182f9180c1bc66bce0a9fb8c