feat: display modal to explain the split
When the kanban is split, users should have a modal whenever they go to
a project with backlog or kanban to explain the split.

If user clicks on Understood button, then Tuleap won't anymore show the
modal again for this user.

Note: tlp-button has been extracted so that the modal on FP pages have
correct styles.

Part of story #33979: have dedicated Backlog service

Change-Id: I68f8bf60d3fd322b6abdc641ba2f1d2cdd6c5db0