feat: no start kanban if kanban service is used
When I go to a project without any scrum and kanban configured, I should
not be able to "start kanban" if the project is using kanban service.

This page will be revamped with a cleaner empty state so we present only
the minimum for now (still hidden by a Feature Flag by the way).

Part of story #33646: access Kanban in a dedicated service

Change-Id: I6b171d50be23b0b938b12752ca36f182216f33a1