Do not use the "In Milestone" criterion value when tracker is no more a backlog tracker
This is part of request #33671 Tracker report "In Milestone" criterion still used when tracker is no more a backlog tracke

Given a tracker defined as a backlog tracker of any planning in the agiledashboard of the project,
and given this tracker has defined in one of its report a value for the "In Milestone" criterion,
then this criterion value is still used to filter artifacts even if the tracker has been removed
as backlog tracker and the criterion is no more visible.

In a following patch, the criterion value will be unset in DB when
removing a tracker as a backlog tracker.

Change-Id: I390f179b80cabce572f2c3dc7c0533eaad20b8ba