refactor: WritingZone is a custom element
part of story #32315 Add support of markdown

This refactoring is needed to simplify the implementation of markdown in
pull-requests comments.

The WritingZone was rendered via UpdateFunctions, and its state was
managed by comments controllers.

The goal is to define it as a custom element having its own controller
managing its internal state. In the next steps, when we'll add a second
tab to the WritingZone, an thus have a more complex management of its
state, then we'll only have to do it in the custom element, and nowhere
else.

How to test:
-- No functional change expected

Change-Id: Ic0072125f2a7d64df1c63df1942b626dc1099ce3