Closes request #33997: Drop support of CVS
Some restrictions regarding the name of users or projects have been kept
"just in case" (and also because they are unlikely to be an issue for
real world use cases).

Note that CVS was the last feature relying on having local unix accounts
for users. The removal of multiple pieces related to the managed of that
will happen in dedicated contributions as it requires a more gentle
touch.

Change-Id: Ia1c6d430e6a581984721c567975b1b1501c12a3f