chore: PlanningBuilder instead of random parameters
No functional change. Unit tests should keep passing.

Notes:
Planning is a good candidate for a Builder object, as it has seven
parameters, most of which are never useful in tests. We never need to
know the name of a planning, except in a single presenter. Therefore,
replacing it with a builder diminishes the chances it will get mocked
and reduces the boilerplate of test setups. We can focus on what's
really important: the milestone and backlog trackers of the planning
object.

part of story #33979 have dedicated Backlog service

Change-Id: I784d716431ebd1eb8490075cf8113e78bb945986