Rename the `Top Planning Backlog`/`Top Backlog` to `Backlog` in A.D backend plugin
Part of #story #33979 have dedicated Backlog service

This is only the first step, it concerns the planning v2 view, the
A.D admin and the `Add to top backlog`  button in the artifact view.
Only the PHP part is translated for now.

This patch take in account the feature flag value, so if you want to see
the change you have to enable the feature flag:
`tuleap config-set feature_flag_activate_split_kanban_for_project 1`

To test:
 - See the Top planning
 - The A.D admin menu
 - The action button `Add Top Backlog` in the artifact view

Change-Id: I67f0dacd19ce97ace8fb165009fc0ddb0f192d82