feat: Syntaxic colorisation in code blocks with specified language
part of story #32315 Add support of markdown

This contributions brings syntaic colorisation in code blocks whose
language is specified. When it is specified, the post processed content
has already its code blocks wrapped in <tlp-syntax-highlighting/> custom
elements, so I only needed to load them in the page.

How to test:
- Write a code block without specifying the language
- Write a code block with a specified language
- Toggle the preview
--> The first code block appears on a grey background, no colorisation
--> The second one also appears on a grey background, and has syntaxic
colorisation based on the specified language

- In file diff, when a line in a code block is too long
--> It is wrapped

Change-Id: I0f34560408210fe388e48957f04678c643b67ebf