feat: remove mention of A.D in import config page
In A.D » administration » import, we remove the mention of A.D so that
we don't have to choose between A.D or backlog.

Took the opportunity to convert the page to Burning Parrot. UX can be
improved further but it is not the aim to redo everything ^^

Note: the link "standard template" does not work but it is already the
case on master. Not my war as they said.

Part of story #33979: have dedicated Backlog service

Change-Id: I46ef1302cf16cf0e3f3908936b048194b1c21be5