fix: WritingZone should not be emptied when user scrolls changes
part of story #32315 Add support of markdown

In CodeMirror editors, the content is rendered when it reaches a given
threshold before entering the view zone.

The problem is that when the NewCommentForm element passes back this
threshold, it is removed from the DOM tree, causing the WritingZone
inside of it to reset completely.

The proposed solution is to keep a trace of the unsaved content typed in
its textarea and reassign it when a new WritingZone is created.

Change-Id: I481438b4dfca04046132f0fb4e52787e47d64605