fix: WritingZone is not focused after connect anymore in file-diff
part of story #32315 Add support of markdown

This issue is provoked by CodeMirror itself. When a gutter is clicked, a
first mousedown event is triggered then its propagation is stopped
immediately, followed by a click event.

Given that the writing zone listens the click event, it detects it as a
click outside of its boundaries, and proceeds to blur the writing zone.

The proposed solution is to make it listen to the mousedown event. In
this case, when a mousedown is dispatched by a gutter click, it won't
reach the WritingZone. When a mousedown event is dispatch by a random
element of the document, the WritingZone reacts.

How to test:
- In the file-diff view, click on a gutter to add a new comment
--> The writing zone is focused

Change-Id: I61f644ec5fa254c7d084b8f3ed31554d31e3e8ab