style: Introduce dark popovers
part of story #32315 Add support of markdown

The markdown cheatsheet popover will be dark. Since these dark popovers
will also be used in the new sidebar, we need to implement them in TLP
first.

Please note that footers are not taken into account. I did not receive
any inputs regarding them. Moreover, we don't have any "button on dark
background" styles yet.

Regarding the differences between the switch-to modal background colors
and the dark popover background colors: It is intended to have a
noticeable color difference between the header and the body. If we
applied the same colors than the switch-to modal, it would be too
subtle, because the modal has a backdrop which improves the contrast
between the two elements, while a popover will almost always be dispayed
on a white background.

How to test:
- In the tlp doc > Fly Over > Popovers, the 4th example demonstrates a dark
  popover.

Change-Id: I28c157e7b6037d396b018171269d5ca34e85380d