Closes request #34021: ForgeUpgrage buckets might not be executed when they come from a not yet enabled plugin
We have the situation in Tuleap 14.12 where the new Kanban plugin is
enabled from the existing AgileDashboard plugin and also has a migration
to run.

Note this is not recommended to enable a plugin this way since the
plugin might need to do some other actions (but it is not the case in
this particular case). Anyway ensuring we do not find new migration
buckets to run after a first pass is not a bad idea as it make the code
more robust.
If one migration fails the whole process exits immediatly so we will not
end up in a infinite loop. If a migration bucket stays pending we will
get a crash due to the recursion, this is not great but this is really
not supposed to happen so a crash is probably not so bad in this case.

In most cases this contribution changes nothing, we just waste a bit of
time checking if we still have buckets to run.

Change-Id: I7e08b8ea9b87d0280d76700c3cc238d3a44d71ce