feat: @tuleap/commonmark-popover to display commonmark cheatsheet
part of story #32315 Add support of markdown

This contribution introduces a new custom element whose role is to
display a commonmark cheatsheet. It is available as a lib, so it can be
reused everywhere we can write commonmark.

I needed to know whether the popover was hidden or shown, so I've made
it dispatch events, like it is already done in tlp-tooltip and
tlp-modal.

How to test:
- In WritingZones headers
--> There is a trigger "Markdown available (?)" on the right end

- hover it
--> It is underlined

- Click it
--> The popover is shown
--> You can scroll it
--> The trigger text appears in tlp-main-color

Change-Id: Ief302e999d36fc329e2f90c8344dcfaa43c7c1e5