feat: Disable Program when Backlog is enabled and feature flag is on
How to test:
- Enable the feature flag: `tuleap config-set
feature_flag_activate_split_kanban_for_project 1`
- Set up a project with no Backlog configuration: create a new project
from tracker template, then go to project admin > Services and enable
the Backlog service
- In the Services admin, you cannot enable the Program service. The
message explains that it is because the Backlog service is also enabled.
- If you disable the Backlog service, you should be allowed to enable
the Program service again.

- If you disable the feature flag, the rules are the same as before: you
are allowed to have AgileDashboard service (for Kanbans) and Program
service at the same time.

Notes:
Program Service will be blocked when Backlog Service is enabled, because
now Project Backlog empty state allows to "start scrum". If someone
enables Program, and then enables Backlog, they can "start scrum", but
they should not be allowed to do that (it was blocked by various Events
for Agile Dashboard). We could just prevent the "start scrum" button,
but it would result in a service that is always empty. It makes no
sense, so we should "just" base the decision on the Backlog service
itself.

The Agile Dashboard service was allowed only for Kanbans, all the rest
was disabled by various Events. Now that the two are split, we should
base the decision on the Backlog Service itself instead of its
configuration.

part of story #33979 have dedicated Backlog service

Change-Id: I5e9cb5fbdf6804d2f1079ce12d7afe246c4076eb