fix: load tooltip on TTM executions
When a test case contains a reference to an artifact, we should display
a tooltip like in the rest of Tuleap. The following situations are
covered by this contribution:
* Test case description
* Step description & expected results (text/html/markdown)
* Test execution comment(*)
* Link to requirement
* Link(s) to (many) bug(s)
* Feedback after link to a bug

(*) Please note that due to current behavior of Tuleap tooltip, there is
a high chance that the tooltip displayed for a xref in the test
execution comment is displayed too low to be readable. This requires a
rewrite of the @tuleap/tooltip library in order to use @floating-ui
(like for TLP popovers) instead of the legacy hand made implementation.
This should be done in a dedicated contribution since other places in
Tuleap can be impacted.
In the meantime, as a workaround, users can expand the comment zone in
order to access to the information.

@tuleap/tooltip changes:
* A method to load only the tooltip on a specific anchor element is
exported in order to be able to have a tooltip individually without
having to give a specific classname. We don't need sparklines in this
context, we should be safe.
* Furthermore the href used to retrieve data is retrieved only when
needed. This is useful in angularjs where the href can be updated after
the load of the component.

Closes request #34027: Artifact tooltip is missing in TTM

Change-Id: Idd90fc4a78d63ce26d106f741ed3464416dbc12b