feat: no moar Agile Dashboard category for widgets
Agile Dashboard category is replaced by two categories Backlog & Kanban.

This is not guarded by a feature flag since we must keep widgets in the
same category between personal and project dashboard to not confuse
people. Since feature flag for the split of kanban depends on the
project, we cannot use it.

Part of story #33979: have dedicated Backlog service

Change-Id: Ib67f446c975c7301929cc4ea9482029a26ee92e6