feat: Forbid activation of Backlog service in Programs
How to test:
- Activate the split: `tuleap config-set
feature_flag_temporarily_deactivate_split_kanban_for_project 0`
- In a project using the Program service, go to project administration.
In the Services part, you can see the "Backlog" service. When you click
on edit, in the modal, the checkbox to enable/disable the service is
disabled with a message. The message explains that the Backlog service
cannot be enabled when the project also uses the Program service.
- If you try to enable the Backlog service anyway (by forging the HTML),
it is rejected with an error message.
- In other projects (for example Team projects), the checkbox is active
and can be changed.
- If you deactivate the split (via the feature flag), the checkbox is
active and can be changed.

- In previously existing Program projects, the Backlog / Agile Dashboard
service should be disabled.
- When you create a new project from the Agile Release Train template,
the Backlog service should be disabled. The Kanban service should be
enabled.

part of story #33979 have dedicated Backlog service

Change-Id: I7f1897c0aa7abc990b42706b985233462a76803a