feat: force split kanban for program
If a project is using Program service, then we should ignore the feature
flag and always split kanban.

This is a tradeoff to manage Backlog/Program services exclusion.

We want that program project (train template) to not have backlog and
have kanban. However if split kanban is deactivated by a feature flag,
then we end up with no agiledashboard and no possibility to access to
Features kanban unless project admin think to activate the
agiledashboard service. To workaround this, every project using Program
service will have split kanban, regardless of the feature flag.

Part of story #33979: have dedicated Backlog service

Change-Id: I3432fe791b9a5488080287dc122c909369eb7479