fix: Do not block activation of Program when Backlog is disabled
How to test:
- Activate the split: `tuleap config-set
feature_flag_temporarily_deactivate_split_kanban_for_project 0`
- In a project using the Backlog service, go to project administration.
In the Services part, you can see the "Program" service. When you click
on edit, in the modal, the checkbox to enable/disable the service is
disabled with a message. The message explains that the Program service
cannot be enabled when the project also uses the Backlog service.
- If you try to enable the Program service anyway (by forging the HTML),
it is rejected with an error message.

- In a project that does not use the Backlog service, the "Program"
service can be enabled, there is no message.

- If you deactivate the split (via the feature flag), you will see the
"Agile Dashboard" service. If Scrum has been started (there are planning
configurations), the "Program" service cannot be enabled (with another
message).

Notes: conditions are hard…

part of story #33979 have dedicated Backlog service

Change-Id: I472b63ddaad8fc20c1bf8a4b4dce090c5ef4c04f