fix: implement request #34337 Do not forbid inactivation of services
How to test:
- Forge your DB and enable both Program and Backlog services in the same
project.
- Go to Project Admin > Services. When you edit Backlog and Program
services, the checkbox should be active and can be changed. There should
be no warning message.
- Deactivate one of them, and edit again. The warning telling you that
"you cannot activate Backlog because Program is active" (and vice-versa)
should be shown again. The checkbox should be disabled.

- No functional change on the activation of Test Management service. It
cannot be activated if the Covered by system type is disabled for the
project.
- No functional change on the activation of Mediawiki standalone
service.

Change-Id: Ie599be775d2230af7f37587679f55a4b13a3bc2d