Close request #34344 Git nested tags to unreferenced tag must directly retrieve initial commit SHA1
Given some nested tags in a Git repository, the current implementation will do a loop on each nested tags before retrieving the initial commit. 
In some cases, this is possible to have a tag that references another tag which is not referenced. 
In that particular case, we have to directly retrieve this initial commit by using the rev-list Git command.

A real example of this is the 6.11-3tag pushed in the Tuleap source code.

Change-Id: I32c821b14bdcc2fb5c32c8e1385b3f6526e2f846