feat: Replace fa-tlp-tracker icon
Now the icon for the tracker service is in a disc.

For artifacts in Heartbeat widget or in SwitchTo modal, the old icon is
kept since it is already presented in a disc and would appear too small
otherwise.

Old icon is also used in breadcrumbs, however there is a specific css
rule to not display those icons (introduced 3 years ago). It is time to
remove those icons in the code, it will be done in a dedicated
changeset.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I8f6f7d6e08d5bbaa0a3f725f5c6fb304f79d9296