fix(mediawiki): create missing users prior to migration
When a 1.23 MediaWiki is imported via the XML import/export mechanism,
the pages are created but not the users (in the MW database). That is
not a problem for MW 1.23 but after migration to 1.35 the pages no
longer show up.

In order to fix that, prior to the migration, we have to re-create the
missing users (that are referenced in the revision table).

In order to manually test, you must have an XML archive of a project
with various users that did edit. Without this patch, you cannot do
import + migration and have the content.

Closes request #32305 Migration from 1.23 doesn't show the pages

Change-Id: I3aa71f8db35fd9d71c7fee436615207b29c0629e