feat: Sidebar is not anymore collapsable
We want to believe (🛸) that collapsing the sidebar is not a feature
really used.

In order to assert that, we remove the possibility for the users to do
that, but we allow platform administrator to reactivate the feature via
a feature flag in case of fire in their support channel about that
removal.

To activate :
tuleap config-set feature_flag_allow_collapse_of_sidebar 1

(default is 0)

Part of story #34002: see promoted trackers in sidebar

Change-Id: I27a514a712281ae3d921f150f314fbb52d9687de