refactor: extract CheckPromotedTrackerConfiguration
For better SRP, the check of feature flag is moved in a dedicated
object.

This is necessary to be able to adapt the trackers promotion it trackers
administration: tooltips and feedbacks should mention that promotion is
for +New dropdown AND sidebar (this will be done in a dedicated
changeset).

No functional changes, the display of trackers in sidebar still depends
on feature flag.

tuleap config-set feature_flag_display_promoted_trackers_in_sidebar 0
tuleap config-set feature_flag_display_promoted_trackers_in_sidebar 1

Part of story #34002: see promoted trackers in sidebar

Change-Id: Ief73e1f23b0ed3f9f32a6e095aec43fd8b865a20