feat: mark current tracker as active in sidebar
When a user go to an artifact (or every pages) of a tracker that is
promoted in sidebar, then the tracker is marked as active in the
sidebar. The service tracker is not marked as active to not have two
active elements at the same time without visual differences.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I97c51917f73e8e79a3c3dcc288d76dffeb963270