feat: + icon is too big in sidebar
The + icon for a promoted item in sidebar should be smaller.

Furthermore we should increase the line-height so that descenders are
not cut. To reproduce, have a promoted tracker like "Bug" or
"Repository": descenders for g or y should not be cut.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I7e855657d651d68b11ad6fb81839ce8556bc0af8