fix: Kanban widget fatal error during project creation
As site admin, create a project based on template Kanban. Go to trackers
» activities » administration » tracker permissions. Only project
members can access to the tracker. Switch this new project as template.

As regular user, create a project based on this new template. You should
not get an Internal Server Error

Note: there is still an issue with permission management in templates:
IMHO when siteadmin decides that a project can be used as template, we
should not check permissions during trackers, kanban, … duplication.
However it is not in the scope of this contribution and should be
discussed and fixed in a dedicated one.

Closes request #34369: Kanban widget fatal error during project creation

Change-Id: I131f8c270c6ead7cf5baac531aafdce0e76d4c31