feat(artifacts): Move artifacts is no more linked to artifact deletion
This is part of request #34032 Move of artifacts should not depend on artifact deletion

The move action is no more regultaed using the artifact deletion limit
(this was confusant for users). Instead, each tracker can define if move
action is allowed or not (allowed by default).

How to test
-----------
* As site administrator define a deletion limit to 0

* As tracker administration, allow artifact move action
* As regular user move an artifact
 => artifact move succeed

* As tracker administration, disallow artifact move action
* As regular user, try to move an artifact
 => artifact move must be forbidden

Note: Some points like XML export/import of the option, the output in
REST full tracker representation and tracker inheritance will be handled
in a dedicated commit.

Change-Id: I2ae9c8157c6afc7c1c340b15c57d7302004f0954