fix: unit tests always pass even if we change conditions
In git #tuleap/stable/9f08593c02e5da12a9ad95c8e90b9cdd5682d5ce unit
tests have been change to take into account the check of split kanban.
However some tests always pass even if we change the conditions.

For example if you comment the instruction
`$event->enablePlanningAdministrationDelegation();` in
`testWhenPlanningAdministrationIsDelegatedItDoesNotCreateFirstScrum`
then the test should fail.

This contribution fixes that so that we can work on a clean state.
Since tests were written before the introduction of split, we consider
that the split is deactivated.

Next step will be to add tests when split is activated.

request #34370: Scrum is initialised even if administrator did not decide to

Change-Id: Ibeafc2bb62642d76e1e9f8147ba71cec76bb6ba6