refactor: extract InProject from InProjectWithoutSidebar
When InProjectWithoutSidebar had been introduced, we only needed a link
to return to the project (this is used by OnlyOffice integration).

Now that we are refactoring the whole `$params` crap, there are places
where we need to store project specific information (the project itself,
the current service "toptab", …). The current contribution is a first
step to reach that goal (eg of possible usage:
`$params->in_project->project`, or
`$params->in_project->current_service`).

No functional change in OnlyOffice integration, the back link is still
displayed in the header.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I83a7ce40d623c83a9cf8949d6b908bf3bc6890db