refactor: Use Option instead of null in Initial changeset
How to test:
- No functional change in the REST API POST /artifacts. You can still
create a new artifact with a parent through the "parent" ket and set
forward links through the "links" key. You can still set forward and
reverse links through the "all_links" key.

Notes:
Using Options instead of nullable should make it easier to tell when a
value is submitted or not. For example, user can omit the "parent" key,
or there could be no artifact links field in the tracker (hence no value
for artifact link field in the changeset values).

part of request #34374 Be able to explicitly add, change or remove
forward and reverse links

Change-Id: I47a1e39c798d90d4e6b9d27974239238217876bc