refactor: active promoted item in HeaderConfiguration
The new parameter that has been introduced for tracker & kanban promotion
is now part of HeaderConfiguration. No functional changes in the
sidebar, current kanban is still marked as active.

This forced us to include new_dropdown_current_context_section as well,
and to discover that there was a bug introduced in another refactoring
two month ago: now when you go on a kanban, there is a section with the
current tracker in the +New dropdown. Before we were taking the id of a
kanban in the url parameters, but now it is part of variables.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I7b2efc35d45b98209abade2dc3e2de59e990daab