fix: Create artifact and add reverse links in a transaction
How to test:
- Using the POST /artifacts REST API, if you create a new artifact and
add a reverse link to an artifact with no Artifact link field or from a
tracker you don't have permission to see, it should cancel the artifact
creation as well as the update of the reverse link. It should not end in
a half-done state.

Implements request #34389 Missing transaction when adding reverse links
to created artifact

Change-Id: I01f40af7a9f3dd5bd47b8c4d7526011fb106954f