fix: No sidebar for onlyoffice integration
Introduced by git #tuleap/stable/1a390ec3a201c480714cd4f9fa75d7a553351450

<mode rant=on>
Once again we build features on a pile of 💩 and when we touch one part,
it breaks another one 😫
</mode>

When we are in a project without sidebar, we should consider that we are
not in a project at all, like it was done before.

Go to document » onlyoffice integration, you should not have a sidebar.
Other pages of the project or siteadmin still have sidebar displayed
correctly.

Part of story #34002: see promoted trackers in sidebar

Change-Id: I065aacae172c6d5d557cc123a0e37b2fcf36701d