fix: Selected open list value not displayed in Artifact Modal
The check performed to know if a given value is selected or not is made against the values ids. Nevertheless, the field value id is casted to string while the selected values ids are assumed to be already strings while it is not the case.

In order to prevent potential bugs because of the number vs string ids, the easier solution is to cast both ids to string.

Closes request #34713 Selected open list value not displayed in Artifact Modal

Change-Id: Ib958fef5d016436fca778d05fb0a8658a59754a7