feat: limit nb of displayed linked projects in sidebar
When the
number of linked projects exceeds the limit of 5, then they are no more
displayed in the sidebar, only in the popover on hover.

Note: the popover is currently barely usable, since it is hidden when we
try to reach it. A dedicated contribution will fix that.

Part of story #34009: see aggregated & parents projects

Change-Id: I90fd2450b3e3459c6a672ec07954c7d52e31c561