chore: add missing tests for new sidebar config
Follow-up to 8b16812d3bca79e07f47c4728a0765e0471b69fd

A test was added to test the default behavior, but not for actual
configuration. This is wrong because the limit could be hardcoded in
LinkedProjects instead of reading the config.

Also remove the snapshots that was not really usefull here.

Part of story #34009: see aggregated & parents projects

Change-Id: Icb565de87c60d6b154f703d10d8c86621161752a