Be able to edit a pull request comment via REST route
Part of story #32316 Update comments

Introduce the new route `PATCH pull_requests_comments/:id` route which
update a global comment of a pull request

To test:
 - Enable the feature flag `tuleap config-set feature_flag_allow_pull_requests_comments_edition 1`
 - launch at least 'make composer && make dev-clear-cache`
 - You can test with API explorer under the `pull_requests_comments` part
 - payload example:
{
  "content": "B3 Touring"
}
=> If the current user tries to edit a comment from another user an error is retruned
=> If the format is not `commonmark` then an error is returned
=> If the user does not have access to project / pull request or git
repository, an error is returned
=> Else the wanted comment is edited and you can see it in the view
=> Nothing is returned by the route for now. It will be done later.

Notes:
To avoid using some `Factory` classes which can do so many stuff (such as build a car and retrieve a
   coffee at the same time), the retrieving of the pull request and
comment are done in dedicated class. Since the review is already long
enough, the refactoring of the `Factory` class will be done in dedicated
commit.

The REST tests will be also done in a dedicated commit

Change-Id: I21d41d142932fdba2919ea490beea6e101e768e5