fix request #34750 Cannot start meilisearch server on el9 when database is on localhost
We need to grant AF_UNIX in addition to AF_INET & AF_INET6.

It's still a bit of protection as we don't allow stuff like AF_PACKET or
anything that could come later one.

How to test
-----------

On el9 environment with systemd and database on server, without this patch
`systemctl start tuleap-meilisearch` crash on database connection.

With the patch start succeed.

Change-Id: Ia4d407fc145a53e8faa5dfc8d8e0241fbac1f96c